+++ model/src/main/java/tech/metavm/object/instance/rest/TreeDTO.java
@@ -1,6 +1,6 @@
 package tech.metavm.object.instance.rest;
 
 public record TreeDTO(
-        long id, long version, int nextNodeId, byte[] bytes
+        long id, long version, long nextNodeId, byte[] bytes
 ) {
 }
